Feature/ZG-168 服务器战斗验证

+++ utils/functions.go
@@ -57,6 +57,6 @@ func BetweenInt32(n, a, b int32) bool {
 }
 
 // Round 四舍五入
-func Round(val float64) int {
-	return int(math.Round((val*10 + 0.1) / 10))
+func Round(val float64) float64 {
+	return math.Round((val*10 + 0.1) / 10)
 }
